Cleaned up priority reporting in normal output and removed it from json output when no high priority commands have been sent. Removed priority statistics reporting for non-read operations since the priority feature is only relevant to reads.

+++ iolog.c
@@ -886,17 +886,17 @@ void flush_samples(FILE *f, void *samples, uint64_t sample_size)
 		s = __get_sample(samples, log_offset, i);
 
 		if (!log_offset) {
-			fprintf(f, "%lu, %" PRId64 ", %u, %u, %u\n",
+			fprintf(f, "%lu, %" PRId64 ", %u, %llu, %u\n",
 					(unsigned long) s->time,
 					s->data.val,
-					io_sample_ddir(s), s->bs, s->priorityBit);
+					io_sample_ddir(s), (unsigned long long) s->bs, s->priorityBit);
 		} else {
 			struct io_sample_offset *so = (void *) s;
 
-			fprintf(f, "%lu, %" PRId64 ", %u, %u, %llu, %u\n",
+			fprintf(f, "%lu, %" PRId64 ", %u, %llu, %llu, %u\n",
 					(unsigned long) s->time,
 					s->data.val,
-					io_sample_ddir(s), s->bs,
+					io_sample_ddir(s), (unsigned long long) s->bs,
 					(unsigned long long) so->offset, s->priorityBit);
 		}
 	}
